Implement rainbow color change hover effect
The variables R, G, and B are randomized values ranging from 0 to
255. These values are used to generate the rgb color. The rainbow
color change occurs if and only if the background color of a square
in the grid is equivalent to empty('').

If a certain square in a grid is already filled with a certain color,
the rgba value of that square is copied on a variable as a string
array. The values were separated through the use of a regular
expression, and the split function.

The color change of the colored square will only occur if the alpha
value of that square is less than 1, then the alpha value gets an
increment of 0.1. Hence the color of the square gets more opaque for
every hover.